requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
gevent
psycopg2